lib: extract out functions from citgm
This PR makes a lot of churn but it is a necessary step to start
writing unti tests for each of the parts of citgm.
